Move scoped storage provider to an isolated file
