+++ owncloudApp/src/main/java/com/owncloud/android/providers/LogsProvider.kt
@@ -23,7 +23,7 @@ import android.content.Context
 import com.owncloud.android.BuildConfig
 import com.owncloud.android.MainApp
 import com.owncloud.android.data.preferences.datasources.implementation.SharedPreferencesProviderImpl
-import com.owncloud.android.data.storage.LocalStorageProvider
+import com.owncloud.android.data.storage.ScopedStorageProvider
 import com.owncloud.android.lib.common.http.LogInterceptor
 import com.owncloud.android.lib.common.utils.LoggingHelper
 import timber.log.Timber
@@ -36,7 +36,7 @@ class LogsProvider(
 
     fun startLogging() {
         val dataFolder = MainApp.dataFolder
-        val localStorageProvider = LocalStorageProvider.ScopedStorageProvider(dataFolder, context)
+        val localStorageProvider = ScopedStorageProvider(dataFolder, context)
 
         // Set folder for store logs
         LoggingHelper.startLogging(
