+++ owncloudData/src/main/java/com/owncloud/android/data/storage/ScopedStorageProvider.kt
@@ -1,0 +1,33 @@
+/**
+ * ownCloud Android client application
+ *
+ * <p>
+ * Copyright (C) 2021 ownCloud GmbH.
+ * <p>
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2,
+ * as published by the Free Software Foundation.
+ * <p>
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ * <p>
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.owncloud.android.data.storage
+
+import android.content.Context
+import timber.log.Timber
+import java.io.File
+import java.util.concurrent.TimeUnit
+import kotlin.system.measureTimeMillis
+
+class ScopedStorageProvider(
+    rootFolderName: String,
+    private val context: Context
+) : LocalStorageProvider(rootFolderName) {
+
+    override fun getPrimaryStorageDirectory(): File = context.filesDir
+}
